# Conflicts:
#	docs/llms/messaging.md
#	src/Headless.Messaging.Core/Headless.Messaging.Core.csproj
#	src/Headless.Messaging.Testing/Internal/RecordingConsumeExecutionPipeline.cs
#	src/Headless.Messaging.Testing/Internal/RecordingTransport.cs
#	src/Headless.Messaging.Testing/MessageObservationStore.cs
#	src/Headless.Messaging.Testing/MessagingTestHarness.cs
#	src/Headless.Messaging.Testing/README.md
#	src/Headless.Messaging.Testing/RecordedMessage.cs
#	src/Headless.Messaging.Testing/TestConsumer.cs
#	tests/Headless.Messaging.Testing.Tests.Unit/RecordingInfrastructureTests.cs

+++ src/Headless.Messaging.Core/Monitoring/StatisticsView.cs
@@ -8,12 +8,12 @@ public class StatisticsView
 
     public int Subscribers { get; set; }
 
-    public int PublishedSucceeded { get; set; }
+    public long PublishedSucceeded { get; set; }
 
-    public int PublishedDelayed { get; set; }
+    public long PublishedDelayed { get; set; }
 
-    public int ReceivedSucceeded { get; set; }
+    public long ReceivedSucceeded { get; set; }
 
-    public int PublishedFailed { get; set; }
-    public int ReceivedFailed { get; set; }
+    public long PublishedFailed { get; set; }
+    public long ReceivedFailed { get; set; }
 }
